Clean-up old code for multiple handlers
