+++ internal/handler/common/client_metadata.go
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package common provides common code for all handlers.
 package common
 
 import (
