+++ internal/handler/commonerrors/error.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package commonerrors providers errors shared by all handlers.
+// Package commonerrors provides handlers errors.
 package commonerrors
 
 import (
